Fix parsing multi-unit video packet
There was an issue with parsing video packets with multiple NAL units. The
parser was just considering it as a single unit, resulting in info loss and
decoder not working properly (e.g. if the video packet contained sps, pps and
idr units, they were all treated as a single sps unit).

That was happening because the bytes array created in `startCode` closure
expression got deallocated after the closure end, so its reference was not valid
anymore, and contained garbage values (e.g. [0,0,0,0] instead of [0,0,0,1]).

Fixing it by retaining the bytes array as a static class member.